validation for username and password using ajax

+++ clash/project/models.py
@@ -27,12 +27,12 @@ class Response(models.Model):
 
 
 class Register(models.Model):  # extended user model
-    user=models.OneToOneField(User,on_delete=models.CASCADE,related_name='user')
-    phone=models.PositiveBigIntegerField(default=0)
+    user=models.OneToOneField(User, on_delete=models.CASCADE, related_name='user')
+    phone=models.IntegerField(default=0)
     level = models.CharField(max_length=15)
     language = models.CharField(max_length=15)
     total_score = models.IntegerField(default=0)
-    quelist = models.TextField(max_length=255,default="[]")
+    quelist = models.TextField(max_length=255, default="[]")
     bool = models.BooleanField(default=True)
